pyproject.toml
[tool.poetry]
name = "gcdp"
version = "0.1.0"
description = "Goal-Conditioned Diffusion Policy"
authors = ["HalvardBariller <h.bariller@hotmail.fr>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
absl-py = "2.1.0"
aiohttp = "3.9.5"
aiosignal = "1.3.1"
antlr4-python3-runtime = "4.9.3"
asciitree = "0.3.3"
async-timeout = "4.0.3"
attrs = "23.2.0"
beautifulsoup4 = "4.12.3"
black = "24.4.2"
build = "1.2.1"
cachecontrol = "0.14.0"
certifi = "2024.6.2"
cffi = "1.16.0"
cfgv = "3.4.0"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
click = "8.1.7"
cloudpickle = "3.0.0"
cmake = "3.29.5"
contourpy = "1.2.1"
crashtest = "0.4.1"
cryptography = "42.0.8"
cycler = "0.12.1"
datasets = "2.19.2"
decorator = "4.4.2"
deepdiff = "7.0.1"
diffusers = "0.27.2"
dill = "0.3.8"
distlib = "0.3.8"
dm-control = "1.0.14"
dm-env = "1.6"
dm-tree = "0.1.8"
docker-pycreds = "0.4.0"
dulwich = "0.21.7"
einops = "0.8.0"
farama-notifications = "0.0.4"
fasteners = "0.19"
fastjsonschema = "2.20.0"
filelock = "3.14.0"
flake8 = "7.1.0"
flake8-black = "0.3.6"
fonttools = "4.53.0"
frozenlist = "1.4.1"
fsspec = "2024.3.1"
gdown = "5.2.0"
gitdb = "4.0.11"
gitpython = "3.1.43"
glfw = "2.7.0"
gym-aloha = "0.1.1"
gym-pusht = "0.1.4"
gymnasium = "0.29.1"
h5py = "3.11.0"
huggingface-hub = "0.23.3"
hydra-core = "1.3.2"
identify = "2.5.36"
idna = "3.7"
imageio = "2.34.1"
imageio-ffmpeg = "0.5.1"
installer = "0.7.0"
jaraco-classes = "3.4.0"
jeepney = "0.8.0"
jinja2 = "3.1.4"
keyring = "24.3.1"
kiwisolver = "1.4.5"
labmaze = "1.0.6"
lazy-loader = "0.4"
llvmlite = "0.42.0"
lxml = "5.2.2"
markupsafe = "2.1.5"
matplotlib = "3.9.0"
mccabe = "0.7.0"
more-itertools = "10.3.0"
moviepy = "1.0.3"
mpmath = "1.3.0"
msgpack = "1.0.8"
mujoco = "2.3.7"
multidict = "6.0.5"
multiprocess = "0.70.16"
mypy-extensions = "1.0.0"
networkx = "3.3"
nodeenv = "1.9.1"
numba = "0.59.1"
numcodecs = "0.12.1"
nvidia-cublas-cu12 = "12.1.3.1"
nvidia-cuda-cupti-cu12 = "12.1.105"
nvidia-cuda-nvrtc-cu12 = "12.1.105"
nvidia-cuda-runtime-cu12 = "12.1.105"
nvidia-cudnn-cu12 = "8.9.2.26"
nvidia-cufft-cu12 = "11.0.2.54"
nvidia-curand-cu12 = "10.3.2.106"
nvidia-cusolver-cu12 = "11.4.5.107"
nvidia-cusparse-cu12 = "12.1.0.106"
nvidia-nccl-cu12 = "2.20.5"
nvidia-nvjitlink-cu12 = "12.5.40"
nvidia-nvtx-cu12 = "12.1.105"
omegaconf = "2.3.0"
opencv-python = "4.10.0.82"
ordered-set = "4.1.0"
packaging = "24.1"
pandas = "2.2.2"
pathspec = "0.12.1"
pillow = "10.3.0"
pkginfo = "1.11.1"
poetry = "1.8.3"
poetry-core = "1.9.0"
poetry-plugin-export = "1.8.0"
pre-commit = "3.7.1"
proglog = "0.1.10"
protobuf = "^5.27.3"
pyarrow = "16.1.0"
pyarrow-hotfix = "0.6"
pyav = "12.1.0"
pycodestyle = "2.12.0"
pycparser = "2.22"
pyflakes = "3.2.0"
pygame = "2.5.2"
pymunk = "6.8.1"
pyopengl = "3.1.7"
pyopengl-accelerate = "3.1.7"
pyparsing = "3.1.2"
pyproject-hooks = "1.1.0"
pysocks = "1.7.1"
python-dateutil = "2.9.0.post0"
pytz = "2024.1"
pyyaml = "6.0.1"
rapidfuzz = "3.9.3"
regex = "2024.5.15"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
safetensors = "0.4.3"
scikit-image = "0.23.2"
scipy = "1.13.1"
secretstorage = "3.3.3"
sentry-sdk = "2.5.1"
setproctitle = "1.3.3"
setuptools = "70.1.0"
shapely = "2.0.4"
shellingham = "1.5.4"
smmap = "5.0.1"
soupsieve = "2.5"
sympy = "1.12.1"
termcolor = "2.4.0"
tifffile = "2024.5.22"
tomli = "2.0.1"
tomlkit = "0.12.5"
torch = "2.3.1"
torchvision = "0.18.1"
tqdm = "4.66.4"
triton = "2.3.1"
trove-classifiers = "2024.5.22"
tzdata = "2024.1"
urllib3 = "2.2.2"
virtualenv = "20.26.2"
wandb = "0.17.1"
xxhash = "3.4.1"
yarl = "1.9.4"
zarr = "2.18.2"


[tool.black]
include = '\.pyi?$'
line-length = 79

[tool.flake8]
extend-ignore = ["E266", "E501", "D102"]
enable-extensions = ["C4"]

[tool.mypy]
ignore_missing_imports=true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


